fix(estimation): address code-review findings on Bayes estimator [#380]
1. fitrx round-trip bug (CONFIRMED): method_to_str writes "bayes" but
   method_from_str had no matching arm, so loading any .fitrx produced by a
   method=bayes fit failed with Corrupt("unknown method \"bayes\""). Add the
   arm + an exhaustive method_str_roundtrips_every_variant test (guards future
   variants too).

2. Partial mu-referencing (altitude): the conjugate theta Gibbs move requires
   every eta to be a non-fixed log mu-ref; otherwise those theta silently fall
   to the slow random-walk block. run_bayes now emits a warning in that case so
   a high R-hat has an actionable cause. Test: run_bayes_warns_on_partial_mu_ref.

3. Reuse: moved the generic MCMC diagnostics (split_rhat, effective_sample_size,
   quantile_sorted + mean/sample_var/autocov) out of estimation/bayes.rs into a
   new shared src/stats/convergence.rs so future samplers/SIR summaries reuse
   one implementation; bayes.rs keeps only the Bayes-specific summarize_param.
   Diagnostic unit tests moved with them.

1166 lib tests pass under --features ci. Refs ferx-core#380.
