+++ src/io/fitrx.rs
@@ -439,6 +439,7 @@ fn method_from_str(s: &str) -> Result<EstimationMethod, FitrxError> {
         "foce_gn_hybrid" => EstimationMethod::FoceGnHybrid,
         "saem" => EstimationMethod::Saem,
         "imp" => EstimationMethod::Imp,
+        "bayes" => EstimationMethod::Bayes,
         _ => return Err(FitrxError::Corrupt(format!("unknown method {:?}", s))),
     })
 }
@@ -1971,6 +1972,24 @@ mod tests {
         let r0 = minimal_fit_result(); // npde_seed defaults to None
         save_fit(&r0, &p, "src\n", &none, SaveFitOptions::default()).unwrap();
         assert_eq!(load_fit(&none).unwrap().fit.npde_seed, None);
+    }
+
+    #[test]
+    fn method_str_roundtrips_every_variant() {
+        // Guards against a new EstimationMethod variant being added to
+        // method_to_str (write) but not method_from_str (read), which silently
+        // breaks .fitrx load for fits produced with that method (e.g. #380's
+        // `bayes`). Exhaustive over the enum so the compiler-style coverage is
+        // enforced at test time.
+        use crate::types::EstimationMethod::*;
+        for m in [Foce, FoceI, FoceGn, FoceGnHybrid, Saem, Imp, Impmap, Bayes] {
+            let s = method_to_str(m);
+            assert_eq!(
+                method_from_str(s).expect("method_from_str must accept method_to_str output"),
+                m,
+                "method round-trip failed for {s:?}"
+            );
+        }
     }
 
     #[test]
