+++ src/stats/mod.rs
@@ -1,3 +1,4 @@
+pub mod convergence;
 pub mod likelihood;
 pub mod npde;
 pub mod residual_error;
